requirements.txt
fastapi==0.111.0
aiohttp==3.9.5
msal==1.29.0
firebase-admin==6.5.0
Pyrebase==3.0.18
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
pyodbc==5.1.0
uvicorn==0.30.1